Changed SheetOverlay ROI behavior
When overlaying SheetView objects the created SheetOverlay object
will inherit the ROI from the component SheetView. If both component
SheetViews have an roi_bounds, it will be inherited from the
SheetView left of the multiply operator. When calling ROI on the
SheetLayer it applies it's inherited ROI to all the layers within
it. This way layers in a SheetOverlay can retain independent
roi_bounds and a ProjectionGrid can overlay the same SheetView
object on all elements in the grid.