Splits the dataviews subpackage out from imagen and fixes imports
This commit introduces the dataviews package, which consists of
data containers for storing data across n-dimensions, along with
plotting code to visualize the views and an IPython extension to
embed the plots within a notebook. The commit also contains
changes to imports within imagen to reflect the addition of the
dataviews subpackage.